Extract MatchdayOrchestrator from AdvanceMatchday HTTP action
Move 700+ lines of domain orchestration logic out of the HTTP layer into
dedicated service classes:

- MatchdayOrchestrator: all matchday advancement logic (batch processing,
  simulation, post-match actions, career mode operations, competition
  progress checks)
- MatchFinalizationService: match finalization and cup tie resolution
  (extracted from FinalizeMatch action)
- MatchdayAdvanceResult DTO: typed result for action-to-service communication

AdvanceMatchday and FinalizeMatch actions are now thin HTTP wrappers that
delegate to services and translate results into redirects.
