+++ app/Modules/Match/Services/MatchFinalizationService.php
@@ -1,0 +1,69 @@
+<?php
+
+namespace App\Modules\Match\Services;
+
+use App\Modules\Match\Events\CupTieResolved;
+use App\Modules\Match\Events\MatchFinalized;
+use App\Models\Competition;
+use App\Models\CupTie;
+use App\Models\Game;
+use App\Models\GameMatch;
+use App\Models\GamePlayer;
+
+class MatchFinalizationService
+{
+    public function __construct(
+        private readonly CupTieResolver $cupTieResolver,
+    ) {}
+
+    /**
+     * Apply all deferred score-dependent side effects for a match.
+     *
+     * Core logic (cup tie resolution) runs here. All other side effects
+     * (standings, GK stats, notifications, prize money, draws) are handled
+     * by listeners on MatchFinalized and CupTieResolved events.
+     */
+    public function finalize(GameMatch $match, Game $game): void
+    {
+        $competition = Competition::find($match->competition_id);
+
+        // 1. Resolve cup tie and dispatch CupTieResolved if applicable
+        if ($match->cup_tie_id !== null) {
+            $this->resolveCupTie($match, $game, $competition);
+        }
+
+        // 2. Dispatch MatchFinalized for standings, GK stats, and notifications
+        MatchFinalized::dispatch($match, $game, $competition);
+
+        // 3. Clear the pending flag
+        $game->update(['pending_finalization_match_id' => null]);
+    }
+
+    private function resolveCupTie(GameMatch $match, Game $game, ?Competition $competition): void
+    {
+        $cupTie = CupTie::with([
+            'firstLegMatch.homeTeam', 'firstLegMatch.awayTeam',
+            'secondLegMatch.homeTeam', 'secondLegMatch.awayTeam',
+        ])->find($match->cup_tie_id);
+
+        if (! $cupTie || $cupTie->completed) {
+            return;
+        }
+
+        // Build players collection for extra time / penalty simulation
+        $allLineupIds = array_merge($match->home_lineup ?? [], $match->away_lineup ?? []);
+        $players = GamePlayer::with('player')->whereIn('id', $allLineupIds)->get();
+        $allPlayers = collect([
+            $match->home_team_id => $players->filter(fn ($p) => $p->team_id === $match->home_team_id),
+            $match->away_team_id => $players->filter(fn ($p) => $p->team_id === $match->away_team_id),
+        ]);
+
+        $winnerId = $this->cupTieResolver->resolve($cupTie, $allPlayers);
+
+        if (! $winnerId) {
+            return;
+        }
+
+        CupTieResolved::dispatch($cupTie, $winnerId, $match, $game, $competition);
+    }
+}
